+++ src/Symfony/Bundle/FrameworkBundle/CacheWarmer/RouterCacheWarmer.php
@@ -11,53 +11,8 @@
 
 namespace Symfony\Bundle\FrameworkBundle\CacheWarmer;
 
-use Psr\Container\ContainerInterface;
-use Symfony\Component\HttpKernel\CacheWarmer\CacheWarmerInterface;
-use Symfony\Component\HttpKernel\CacheWarmer\WarmableInterface;
-use Symfony\Component\Routing\RouterInterface;
-use Symfony\Contracts\Service\ServiceSubscriberInterface;
+use Symfony\Component\Routing\CacheWarmer\RouterCacheWarmer as BaseRouterCacheWarmer;
 
-/**
- * Generates the router matcher and generator classes.
- *
- * @author Fabien Potencier <fabien@symfony.com>
- *
- * @final
- */
-class RouterCacheWarmer implements CacheWarmerInterface, ServiceSubscriberInterface
-{
-    /**
-     * As this cache warmer is optional, dependencies should be lazy-loaded, that's why a container should be injected.
-     */
-    public function __construct(
-        private ContainerInterface $container,
-    ) {
-    }
+trigger_deprecation('symfony/framework-bundle', '8.2', 'The "%s" class is deprecated, use "%s" instead.', 'Symfony\Bundle\FrameworkBundle\CacheWarmer\RouterCacheWarmer', BaseRouterCacheWarmer::class);
 
-    public function warmUp(string $cacheDir, ?string $buildDir = null): array
-    {
-        if (!$buildDir) {
-            return [];
-        }
-
-        $router = $this->container->get('router');
-
-        if ($router instanceof WarmableInterface) {
-            return $router->warmUp($cacheDir, $buildDir);
-        }
-
-        throw new \LogicException(\sprintf('The router "%s" cannot be warmed up because it does not implement "%s".', get_debug_type($router), WarmableInterface::class));
-    }
-
-    public function isOptional(): bool
-    {
-        return true;
-    }
-
-    public static function getSubscribedServices(): array
-    {
-        return [
-            'router' => RouterInterface::class,
-        ];
-    }
-}
+class_alias(BaseRouterCacheWarmer::class, 'Symfony\Bundle\FrameworkBundle\CacheWarmer\RouterCacheWarmer');