+++ src/Symfony/Bundle/FrameworkBundle/Routing/AttributeRouteControllerLoader.php
@@ -11,41 +11,13 @@
 
 namespace Symfony\Bundle\FrameworkBundle\Routing;
 
-use Symfony\Component\Routing\Loader\AttributeClassLoader;
-use Symfony\Component\Routing\Route;
+use Symfony\Component\Routing\Loader\AttributeRouteControllerLoader as BaseAttributeRouteControllerLoader;
+
+trigger_deprecation('symfony/framework-bundle', '8.2', 'The "%s" class is deprecated, use "%s" instead.', AttributeRouteControllerLoader::class, BaseAttributeRouteControllerLoader::class);
 
 /**
- * AttributeRouteControllerLoader is an implementation of AttributeClassLoader
- * that sets the '_controller' default based on the class and method names.
- *
- * @author Fabien Potencier <fabien@symfony.com>
- * @author Alexandre Daubois <alex.daubois@gmail.com>
+ * @deprecated since Symfony 8.2, use Symfony\Component\Routing\Loader\AttributeRouteControllerLoader instead
  */
-class AttributeRouteControllerLoader extends AttributeClassLoader
+class AttributeRouteControllerLoader extends BaseAttributeRouteControllerLoader
 {
-    /**
-     * Configures the _controller default parameter of a given Route instance.
-     */
-    protected function configureRoute(Route $route, \ReflectionClass $class, \ReflectionMethod $method, object $attr): void
-    {
-        if ('__invoke' === $method->getName()) {
-            $route->setDefault('_controller', $class->getName());
-        } else {
-            $route->setDefault('_controller', $class->getName().'::'.$method->getName());
-        }
-    }
-
-    /**
-     * Makes the default route name more sane by removing common keywords.
-     */
-    protected function getDefaultRouteName(\ReflectionClass $class, \ReflectionMethod $method): string
-    {
-        $name = preg_replace('/(bundle|controller)_/', '_', parent::getDefaultRouteName($class, $method));
-
-        if (str_ends_with($method->name, 'Action') || str_ends_with($method->name, '_action')) {
-            $name = preg_replace('/action(_\d+)?$/', '\\1', $name);
-        }
-
-        return str_replace('__', '_', $name);
-    }
 }