+++ src/Symfony/Component/Translation/DependencyInjection/RemoveMissingDependenciesPass.php
@@ -13,7 +13,6 @@ namespace Symfony\Component\Translation\DependencyInjection;
 
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\Translation\TranslationBundle;
 
 /**
  * Drops the services that the container cannot wire, because what they depend on is missing.
@@ -29,15 +28,21 @@ class RemoveMissingDependenciesPass implements CompilerPassInterface
             $container->removeDefinition('translator.data_collector');
         }
 
-        if (!$container->hasParameter(TranslationBundle::PROVIDER_LOCALES_PARAMETER)) {
+        if (!$container->hasParameter('.translator.provider_locales')) {
             return;
         }
 
         // the enabled locales come from another extension, so they can only be read now
         $enabledLocales = $container->hasParameter('kernel.enabled_locales') ? $container->getParameter('kernel.enabled_locales') : [];
-        $locales = array_values(array_unique([...$enabledLocales, ...$container->getParameter(TranslationBundle::PROVIDER_LOCALES_PARAMETER)]));
+        $locales = array_values(array_unique([...$enabledLocales, ...$container->getParameter('.translator.provider_locales')]));
+        $container->getParameterBag()->remove('.translator.provider_locales');
 
-        $container->setParameter(TranslationBundle::PROVIDER_LOCALES_PARAMETER, $locales);
         $container->getDefinition('translation.provider_collection_factory')->replaceArgument(1, $locales);
+
+        foreach (['console.command.translation_pull' => 5, 'console.command.translation_push' => 3] as $id => $argument) {
+            if ($container->hasDefinition($id)) {
+                $container->getDefinition($id)->replaceArgument($argument, $locales);
+            }
+        }
     }
 }